CLI: remove -f, replace -s SAT with -a MIN -b MAX
We now always output colors so there's no need for -f. It was not
useful because the only thing we do is add colors, and if someone
doesn't want to use colors then they can use plain cat instead.

-s SAT was deriving a minimum R/G/B value from a "saturation" value, but
it was not real saturation, and also it only allowed changing the
minimum, which may not be enough e.g. on light background.

-a MIN -b MAX allow full control of the limits, and thus better.

Also, small refactor of the help messages.