output: reset only foreground color
We modify only the foreground color, so that's the only thing we need
to restore/reset once we're done.

Also, there's no need for reset at the end of each line, so remove it.

Also, set the fg color for tab eventhough it's non-printable, as it may
look nicer in selection (depending on terminal).