keep escape sequences unless -e is used
This is nicer for inputs which include their own formatting,
e.g. bold or background colors or positioning now remain unmodified.
The previous behavior of stripping esc seq can be activated with -e .

This requires us to skip esc sequences while coloring the input, so
we now do that programatically rather than pre-processing with sed.