identify ESC(? and ESC)? for rom select
Previously, after ESC we either entered CSI mode on '[', or exited esc
mode on any other char.

Now we allow 1st char after ESC to also be '(' or ')' which extends the
"1st char after ESC" state, and allows e.g. ESC(B for char rom
selection. This is good enough, and does not pretend to be complete.

E.g. this is used by wttr.in graphs, like `curl v2d.wttr.in/NewYork'.