error-exit if FILE is not found and/or "od" failed
Previously, because by default a pipe exits with the last command exit
code, errors in `cat' or `od' were not detected, resulting in empty
output and success exit (though typically with a message to stderr).

Now we error on such cases using custom pipefail-like behavior. Only
success/failure are detected (not propagating the actual error code).

We don't bother with native pipefail where possible, even if the POSIX
202X draft specifies it, because it requires more code to support both,
and additional subshell to detect. Our solution is small and good
enough, especially for an unimportant utility like shlolcat.