rand: avoid erronous edge cases
In rand(), $sd (seed) is `cksum' output, which POSIX defines as:
  printf("%u %d\n", crc32, in_count);  // when FILE is stdin

Previously the output value $rnd was calculated after removing the
last 4 chars from $sd, however, $sd could theoretically be e.g. "1 1",
at which case nothing is removed, and the calculation would error.

Now we remove the space and anything which follows, and left strictly
with the crc32 value.

We also remove the last char of the crc and prepend 1 to ensure that
the value fits in signed 32 bits value (POSIX sh arithmetic exprs.).