useless cat #3142
od can handle stdin/files same as cat, and with one less process.

The behavior is now fully equivalent to pipefail, though we still exit
with 1 and print usage if od fails - to be improved later.

We no longer process the octals delayed-by-1, because now we can
distinguish an octal read (succees) from exit-code read (EOF-fail).