flush regularly regatdless of newlines
Previously we flushed only on newline, which means we could buffer
a lot for long lines, or worse - if a file is without newlines.

Now we flush every two od lines, which is typically 32 octals (and up
to about 800 bytes at the buffer when using true colors). This is likely
a bit more frequently than before on average, and as a result shells
without builtin printf will run shlolcat slightly slower - that's OK.

This commit doesn't really affect our behavior with line-buffered input
because od buffers blocks anyway (16 octals typically).

If we wanted unbuffered, then we should change the flush to every od
line, and also request od to use blocks of 1 byte (-w1), but that's
a non-standard option, so we're constrained by od block buffering.